Add duplicate bag as ETS option (#83)
Big refactor. Besides of fixing logic errors, it adds `duplicate_bag`
ETS table type.
This has a lot of unnecessary code due to the fact how we store tuples
in the hashtable.

Currently, instead of storing tuples, we store list that we prepend
elements on insertion.
This is unfortunate, because some operations need to return nodes in
insertion order (so we need to reverse), and some needs to filter them.

For example:
- `ets:delete_object` needs to first lookup nodes that are stored, then
filter them and them reinsert to remove matching nodes.
- `ets:take` needs to reverse list.
- `ets:insert` needs to make a full copy of the stored nodes, create a
new heap and store them back with prepended term.
- `ets:lookup` and `ets:lookup_element` needs to map results and reverse
the list.

The better solution would be to store them as individual nodes (in
reverse insertion order, i.e. last inserted is first) but then we need
to return some iterator from hashtable since we'll be building list as
we go.

This also necessarily create a lot of collisions so we need to store
them in linked list. Delete suffers in that case, because we need to
iterate over full linked list instead of deleting entire chain (some
other nodes with collisions may be stored there).
The alternative is disallow storing elements with different key but then
we quickly run out of buckets.

All in all, we need a better datastructure with low memory footprint to
store the nodes that can duplicate. It will most likely simplify ETS
code as well.

Reviewers: @TheSobkiewicz, @mat-hek 


These changes are made under both the "Apache 2.0" and the "GNU Lesser
General
Public License 2.1 or later" license terms (dual license).

SPDX-License-Identifier: Apache-2.0 OR LGPL-2.1-or-later

---------

Signed-off-by: Jakub Gonet <jakub.gonet@swmansion.com>