Avoid using memory_ensure_free_with_roots
Only allocate the result tuple when needed, so there's no term to
preserve between GCs.

Fixes a problem with `Unexpected term: 0x2b` caused by allocating a 2-tuple without initializing it.
This meant the 2 words had GCed item (containing moved marker `0x2B`) or garbage.
When calling `memory_ensure_free_with_roots`, a shallow copy of that tuple (it was the root here)
attempted to move that garbage and was failing.