Batch JS calls (#140)
## In previous episode

- `:emscripten.run_remote_object_script_fn(js_fn, opts) ->
RemoteObject.t()`
- `:emscripten.from_remote_object(remote_refs, :key | :value) -> term()`

This API had some shortcomings (apart from terrible naming) – they
worked on individual objects, which created a lot of overhead when
trying to keep track of multiple JS values.
JS code was expected to be a string with a function taking Wasm module
and generated key (used to keep identity for value returned from JS
function) and returning any object or undefined if tracking wasn't
necessary.

Additionally, we supported similar options like in `run_js` but only in
some combinations, as JS code was necessarily run on main thread only.
The implementation was strongly tied to remoteObject property in Wasm
module and did tracking inside compiled JS code.

## New API

New API has the following surface:
- `:emscripten.run_script_tracked(js_code) -> [TrackedObject.t()]`
- `:emscripten.get_tracked(tracked_refs, :key | :value) -> [{:ok,
term()} | {:error, term()}]`

It renames `RemoteObject` to `TrackedObject` which better reflects
intent.
They use JS function callbacks on Wasm module which allows to overwrite
entire implementation in downstream libraries (e.g. Popcorn).
Requirement for `run_script_tracked` is to work on list of numeric keys
that identify values returned from eval'd script. Requirement for
`get_tracked` is to work on values that are strings.

## JS code evaluation

We require passed `js_code` to be a string of JS code which returns an
array or undefined. It's always run on main thread and doesn't take
options.
This code is passed to `onRunTrackedJs` which is required to evaluate
and return a list of keys that track values from JS array.
`onRunTrackedJs` generates the keys by using exported to Wasm
`get_next_tracked_object_key` function.

> [!NOTE]
> Returned refs aren't required to have 1:1 correspondence with values
returned from JS function.

## Value retrieval
To retrieve a value, list of refs is passed to the `get_tracked` which
calls `onGetTrackedObjects` on Wasm module. It's required to return list
of strings. If any returned value in this list isn't a string (undefined
maps to "key not found", others map to "invalid returned value") it's
reported as an error. All returned values are treated independently.

This allows to retrieve multiple keys at once and splitting the output
into valid and invalid sets of values.

> [!NOTE]
> Order is preserved in list of values returned from `get_tracked` but
it doesn't currently have key included. For non-1:1 mapping between keys
and values this can be worked around by pushing additional values at the
start or end of the returned list.

## Improvements

- `get_next_tracked_object_key` should return range of values instead of
single value.

## License agreement

These changes are made under both the "Apache 2.0" and the "GNU Lesser
General
Public License 2.1 or later" license terms (dual license).

SPDX-License-Identifier: Apache-2.0 OR LGPL-2.1-or-later

---------

Signed-off-by: Jakub Gonet <jakub.gonet@swmansion.com>