wasm: add get_next_tracked_object_key
It's a field in emscripten platform struct. It could be a callback entirely in JS but we use threads.
Threads are emulated via webworkers which have their own contexts and variables aren't shared with each other.
This means that we would lose uniqueness of the keys.

This commit also changes output format from file loaded and ran at <script src=...> to ES6 classes
which have greater control over initialization of Wasm.

Signed-off-by: Jakub Gonet <jakub.gonet@swmansion.com>